requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.1
chromadb==0.4.22
groq==0.4.2
pypdf==4.0.1
sentence-transformers==2.3.1
rank-bm25==0.2.2
python-dotenv==1.0.0